tidy-up test_molecular and test_molfeatures
